requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
matplotlib
seaborn
reportlab
fpdf
python-dotenv
openai